+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - Function handles (aggregation and computed attributes functions) are shown after resizing nodes ([#115](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/115))
 - Edges were sometimes inconsistently rendered as unmarked ([#119](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/119))
 
+### Internals
+
+- Fix: Sensitivity analysis test cases occasionally fail ([#111](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/111))
+
 ## [1.2.2](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.2.1...v1.2.2) - 2023-03-20
 
 ### Fixes and Improvements